requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
sqlalchemy==2.0.41
pydantic==2.12.2
python-dotenv==1.1.1
openai==2.3.0
httpx==0.28.1
starlette==0.48.0
pytest
pytest-cov
